Add writeback addressing modes for shift intructions
Because the shift operations can read from memory and then write back to
the same address, I cannot use the normal addressing modes, since both
the load and the store (used to read the operand) move the PC, the store
happening after the load will cause the operand to be misread.

This unfortunately means I have to create special addressing modes just
for these operations. They work by simply avoiding bumping the PC when
the operand is loaded.